requirements.txt
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
infinity==1.5
intervals==0.9.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
pillow==10.3.0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
typing_extensions==4.11.0
validators==0.28.1
Werkzeug==3.0.3
WTForms==3.1.2
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
